Rearrange force/force model files
